backend/myapp/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
APScheduler==3.10.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
distlib==0.3.6
Django==4.1.5
django-cors-headers==4.4.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
filelock==3.9.0
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.22.0
packaging==23.0
pandas==1.3.5
pbr==5.11.1
platformdirs==2.6.2
pluggy==1.0.0
psycopg2==2.9.5
py4j==0.10.9.5
pycparser==2.21
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
virtualenv==20.17.1
Werkzeug==2.2.2
yarl==1.8.2
